split explorer into a lib and a bin (jexplorer)
